requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0